# Vectorize global memory write in matmul schedule