--- a/test/test_gpu_matmul_sass.cpp
+++ b/test/test_gpu_matmul_sass.cpp
@@ -69,8 +69,6 @@ sass::Container getSASSFor(
   params.double_buffer_options.smem_double_buffer_stage = 4;
   scheduleMatmul(&fusion, params);
 
-  fusion.printTransforms();
-
   at::manual_seed(0);
   auto inputs = fp16MatmulAtInput(M, N, K, layout);